requirements.txt
requests
google-generativeai
groq
faiss-cpu
sentence-transformers
